feat: Add weight-based product support
This commit introduces support for products sold by weight or volume,
requiring a shift from integer to decimal quantities across the application.

- **Data Model:**
  - `Produit`, `LigneVente`, `CartItemDto`, and `MouvementStock` entities/DTOs
    have been updated to use `BigDecimal` for quantity fields instead of `int`.
  - A `uniteDeVente` (unit of sale) field has been added to the `Produit` entity
    and `ProduitDto` to specify units like "PIECE", "KG", or "LITRE".

- **Backend:**
  - `StockService`, `ProduitService`, `DashboardController`, and `RapportController`
    have been updated to perform calculations and comparisons using `BigDecimal`
    arithmetic.
  - Repositories (`ProduitRepository`) have been updated to accept `BigDecimal`
    parameters.
  - Unit tests (`ProduitServiceTest`) were updated to reflect the data model changes.

- **Frontend:**
  - The single and batch product forms (`produit-form.html`, `produits.html`)
    now include a dropdown for `uniteDeVente` and allow decimal inputs for quantity.
  - The cashier interface (`caissier.html`) has been updated to handle
    decimal quantities when adding products to the cart and calculating totals.

+++ src/main/java/com/example/boutique/dto/CartItemDto.java
@@ -1,8 +1,10 @@
 package com.example.boutique.dto;
 
+import java.math.BigDecimal;
+
 public class CartItemDto {
     private Long id;
-    private int quantity;
+    private BigDecimal quantity;
 
     public Long getId() {
         return id;
@@ -12,11 +14,11 @@ public class CartItemDto {
         this.id = id;
     }
 
-    public int getQuantity() {
+    public BigDecimal getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(BigDecimal quantity) {
         this.quantity = quantity;
     }
 }
